function/requirements.txt
azure-functions
azureml-opendatasets==1.0.69
azureml-sdk[automl]==1.0.69
pandas==0.23.4
scikit-learn==0.20.3